move Paywall init logic to paywall-app
