+++ paywall-app/src/pages/_app.js
@@ -4,7 +4,6 @@ import React from 'react'
 class UnlockApp extends App {
   render() {
     const { Component, pageProps } = this.props
-
     return (
       <>
         <Component {...pageProps} />
