+++ paywall-app/src/pages/home.js
@@ -1,6 +1,13 @@
 import Link from 'next/link'
+import dynamic from 'next/dynamic'
 import React from 'react'
+import UnlockPaywall from '../components/UnlockPaywall'
 
 export default function Home() {
-  return <Link href="https://docs.unlock-protocol.com/tools/paywall">Docs</Link>
+  return (
+    <div>
+      <Link href="https://docs.unlock-protocol.com/tools/paywall">Docs</Link>
+      <UnlockPaywall />
+    </div>
+  )
 }
